Another variation of "last MSG" bug fixed
GOD DAMNIT WHY DOESNT THIS STOP. This time, when a IF or ELSE jumps to
the second to last line (when the last lime is END IF), and this second
to last is a MSG command, AND this is the first script to be activated
in the game, THEN you are stuck in a endless loop. Now fixed with
adjusted start parameters for the GUI class. Urgh.